# api/api.yaml
# TODO: Add 403 and 401 responses?
openapi: 3.0.3
info:
  title: Quote API
  description: Quote API
  version: 0.1.0
paths:
  /quote/random:
    get:
      summary: Get a random quote
      operationId: v1RandomQuote
      parameters:
        - in: query
          name: author
          schema:
            type: string
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
      responses:
        '303':
          description: Link to a random quote
          headers:
            Location:
              schema:
                type: string
        '404':
          description: No quote found with given parameters
  # TODO: Return some indicator of whether the new status is liked or not liked?
  /quote/{id}/like/toggle:
    post:
      summary:  Toggle like on a quote
      operationId: v1ToggleQuoteLike
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Like successfully toggled
        '404':
          description: Quote not found
  /quote/{id}:
    get:
      summary: Get a quote by id
      operationId: v1GetQuote
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: The requested quote
          content:
            application/json:
              schema:
                $ref: quote.yaml#/Quote
        '404':
          description: The quote wasn't found
    delete:
      summary: Delete a quote by id
      operationId: v1DeleteQuote
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: The quote was successfully deletes
        '404':
          description: The quote wasn't found
    put:
      summary: Edit or create a new quote
      operationId: v1UpdateQuote
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        description: Quote to add
        required: true
        content:
          application/json:
            schema:
              $ref: quote.yaml#/Quote
      responses:
        '201':
          description: A new quote was created
        '204':
          description: The quote was modified
  /quote:
    post:
      summary: Add a new quote
      operationId: v1AddQuote
      requestBody:
        description: Quote to add
        required: true
        content:
          application/json:
            schema:
              $ref: quote.yaml#/Quote
      responses:
        '201':
          description: Created
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: validation-error.yaml#/ValidationError